*** fixed major shortcoming in playing back ultra long recordings where the int position was wrapping around to negative file positions. Now the file position is a long and all methods have been refactored in all classes that use this interface.  verified that a 5 days recording can be handled now.
git-svn-id: https://svn.code.sf.net/p/jaer/code/trunk/host/java@3339 b7f4320f-462c-0410-a916-d9f35bb82d52

+++ src/ch/unizh/ini/jaer/projects/opticalflow/graphics/MotionViewer.java
@@ -777,11 +777,11 @@ public class MotionViewer extends javax.swing.JFrame implements PropertyChangeLi
             motionInputStream.mark();
         }
 
-        public int position() {
+        public long position() {
             return motionInputStream.position();
         }
 
-        public void position(int event) {
+        public void position(long event) {
             motionInputStream.position(event);
         }
